pyproject.toml
[project]
name = "bigdata-job-market"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "cloudscraper>=1.2.71",
    "confluent-kafka==2.4.0",
    "fastapi==0.111.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.3",
    "pyspark==3.5.1",
    "python-dotenv==1.0.1",
    "redis==5.0.4",
    "requests>=2.31.0",
    "uvicorn==0.30.1",
]
